check that switches are on enums
This is imprecise. Really we should be looking up the enum and making
sure that it is the apply_result_t enum for some table. There's no way
to do this nicely right now besides just looking at the name of the
enum, which feels like a hack, so I'm just checking that it's an enum.