Initial support for unit testing
Refactor Common to eliminate p4file and p4filecontents arguments. We
can just pass a single string that is either a file name or the file
contents. And on the Web, we can just use a different pre-processor
argument to the functor that processes the string rather than trying
to load the file.

Added a test directory and set up initial support for parser-only
tests, with one simple example to get us going.