Fix bug in order of arguments to type_param_arg
We should check that the argument type is a subtype of the parameter
type, but the code had it the other way around!

Among other things, this is needed to allow the following code to type
check:
```
extern void f<T>(in tuple<T> x);

control c() {
    apply {
        f<bit<32>>({ 32w2 });
    }
}
```
where the parameter type is `tuple<bit<32>>` and the argument type is
a singleton list with a `bit<32>` element.