Add back earlier stubs for lvalue handling.
The method of passing around locations on the heap as lvalues does not
scale, because it means we need to implement some kind of deep copy
when, for instance, handling a header-to-header assignment. This is
especially hard in Coq, where all fixpoints need to terminate and such a
property cannot be inferred because of the indirection.