update git version
the last step fails with the following error. Some issues state that it's because of the git version, which currently is 2.7.4. So trying this first.
Run git add .
fatal: Not a git repository (or any parent up to mount point /__w)
Stopping at filesystem boundary (GIT_DISCOVERY_ACROSS_FILESYSTEM not set).
Error: Process completed with exit code 128.